+++ .gitignore
@@ -119,4 +119,3 @@ dist
 .yarn/build-state.yml
 .yarn/install-state.gz
 .pnp.*
-/uploads/
