+++ package.json
@@ -9,18 +9,11 @@
   "author": "Alexander Samiylenko <65536sam@gmail.com>",
   "license": "ISC",
   "dependencies": {
-    "@jsonlines/core": "^1.0.2",
-    "JSONStream": "^1.3.5",
-    "bfj": "^7.0.2",
-    "csvtojson": "^2.0.10",
-    "event-stream": "^4.0.1",
     "express": "^4.17.1",
     "find": "^0.3.0",
     "jsonlines": "^0.1.1",
     "papaparse": "^5.3.0",
     "promisify": "0.0.3",
-    "split2": "^3.2.2",
-    "through2": "^4.0.2",
     "uuid": "^8.3.1"
   },
   "devDependencies": {
