+++ src/controllers/uploadJsonController.js
@@ -1,6 +1,5 @@
 const { uploadCsv } = require('../utils/csvToJson');
-// const { optimizeJSON, transformStream } = require('../utils/jsonOptimizer');
-const { optimizeJSON } = require('../utils/jsonOptimizer2');
+const { optimizeJSON } = require('../utils/jsonOptimizer');
 const path = require('path');
 const fs = require('fs');
 
@@ -18,7 +17,6 @@ async function uploadJson(request, response) {
   response.setHeader('Content-Type', 'application/json');
   response.statusCode = 200;
   response.end(JSON.stringify({ status: 'ok' }));
-  return;
 }
 
 function getFilesList(request, response) {
@@ -32,8 +30,8 @@ function getFilesList(request, response) {
 function optimize(request, response) {
   const filename = path.basename(request.url);
   response.setHeader('Content-Type', 'application/json');
-  optimizeJSON(filename, response);
   response.statusCode = 202;
+  optimizeJSON(filename, response);
 }
 
 module.exports = {
