+++ uploads/08324c88-f6cd-407c-b839-43e186187fde.json
@@ -1,0 +1,10 @@
+{"type":"socks","color":"red","quantity":"8","price":"10","isPair":"true"}
+{"type":"socks","color":"red","quantity":"12","price":"10","isPair":"true"}
+{"type":"socks","color":"purple","quantity":"8","price":"2","isPair":"true"}
+{"type":"jeans","color":"navy","quantity":"4","price":"35","isPair":"false"}
+{"type":"gloves","color":"black","quantity":"5","price":"15","isPair":"true"}
+{"type":"gloves","color":"black","quantity":"10","price":"15","isPair":"true"}
+{"type":"hat","color":"lime","quantity":"3","price":"7","isPair":"false"}
+{"type":"socks","color":"red","quantity":"14","price":"2","isPair":"true"}
+{"type":"socks","color":"red","quantity":"6","price":"10","isPair":"true"}
+{"type":"socks","color":"red","quantity":"16","price":"2","isPair":"true"}
