+++ uploads/optimized/08324c88-f6cd-407c-b839-43e186187fde.json
@@ -1,0 +1,6 @@
+{"type":"socks","color":"red","quantity":"26","price":"10","isPair":"true"}
+{"type":"socks","color":"purple","quantity":"8","price":"2","isPair":"true"}
+{"type":"jeans","color":"navy","quantity":"4","price":"35","isPair":"false"}
+{"type":"gloves","color":"black","quantity":"15","price":"15","isPair":"true"}
+{"type":"hat","color":"lime","quantity":"3","price":"7","isPair":"false"}
+{"type":"socks","color":"red","quantity":"30","price":"2","isPair":"true"}
