+++ uploads/optimized/b0cb4c8f-d02f-440f-8a05-4fe806073fef.json
@@ -1,0 +1,6 @@
+{"type":"socks","color":"red","quantity":"3328","price":"10","isPair":"true"}
+{"type":"socks","color":"purple","quantity":"1024","price":"2","isPair":"true"}
+{"type":"jeans","color":"navy","quantity":"512","price":"35","isPair":"false"}
+{"type":"gloves","color":"black","quantity":"1920","price":"15","isPair":"true"}
+{"type":"hat","color":"lime","quantity":"384","price":"7","isPair":"false"}
+{"type":"socks","color":"red","quantity":"3840","price":"2","isPair":"true"}
